Reformatted the names of the Folders and Files
After learning the naming and folder conventions for Java and Cucumber,
I reformatted the items for programming/testing - including creating a
'old' folder for areas and populating them with the previously incorrectly
named files. This updates the packages for Java and Resources as well as
the names of the files used.

Added in @tags for the TopMenuMainLinks, TopMenuSubLinks, FooterLinks
and OnPageLinks features

Areas Updated:
   Java
 - Runners package and runner file
 - StepDev
   Resources
 - FooterLinks
 - OnPageLinks
 - TopMenuMainLinks
 - TopMenuSubLinks